Further work on uniforming decay constants.
There are now two decay constants:
- decay_const: decay constant (in seconds) to accumulate all kind
  of recursive statistics (hos, envelope, RMS, MB-filter).
- rosenberger_decay_const: optional decay constant for the Rosenberger
  algorithm (defaults to decay_const, if not specified).

The "C" parameter is computed as 1./decay_nsmps, where
decay_nsmps is the decay constant converted to samples.

This is simpler than Pierre's formula:

    C = exp(log(0.05) / decay_nsmps)

but consistent to the definition of C given in Lomax et al. (2012),
eq. (1).

This commit also updates several variable names, for better readability
and fixes code style here and there.