MBfilter_CF(): fix for single-component analysis.
Allows any component (and not just 'Z') to be used in single-component
analysis. No assumption is made on component name.

However, three-component analysis makes the assumption that components
are named as 'N', 'E', 'Z'. This should be fixed and generalized in
a future commit.

This commit also removes a useless line in GaussConv().