Optimize for memory usage - part 2.
The biggest change in this commit is that LocalCC no more calculates
both the smoothed and the unsmoothed CC. This is now controlled, at the
level in the function "localCCr()" by the parameter "sigma" (a None
value means no smoothing).

NOTE: The code must be recompiled, since a C function is modified.

TODO: The pure python implementation should be modified accordingly,
or deleted!

LocalCC is no more a class, but a simple function. It is now clearer
what it does and it will maybe save some overhead time related to
object construction/destruction (not sure, though).

Using the "memory_profiler" module I found that the use of "np.where()"
for determining max lag and time from cc was wasting quite a lot of
memory. I replaced it with "np.argmax()" which is memory-efficient.

There is some other small imprevement here and there, which I cannot
remember: just look at the diff, if you're curious :).

Finally, this commit fixes issue #2:
https://bitbucket.org/npoiata/backproj/issue/2