stack_grid as NLL grid. Save it in NLL file format
'stack_grid' is now an NLL grid, which is convenient, since it
stores all the information on grid size, origin, and geographic
transformation.

When we save it to disk (config.save_projGRID), it is now stored
as NLL file (.hdr + .buf) and no more as pickle file.

Also, this simplifies the interface of 'bp_plot()', wich doesn't
need anymore 'grid1' in order to get grid information.

Other modifications in this commit include code replacement with
functionally equivalent, but more idiomatiac, python expressions
(especially in 'bp_plot()').