Config option for HOS sigma parameter.
This parameter ('hos_sigma') sets a sort of minimum detection
threshold for HOS characteristic function.
Its unities are trace amplitude.
If not specified, default value is 0.001.

The parameter in config file must be a comma-separated
list, even for a single value.

If more values are provided, they will be attributed to
the stations (defined by "stations" option), following their
order, e.g.:

  stations = PB02, PB08, PB12
  hos_sigma = 500., 1000., 600.

If there are less values of 'hos_sigma' than 'stations',
the last value is repeated to fill up the station list, e.g.:

  stations = PB02, PB08, PB12
  hos_sigma = 500., 1000.

is equivalent to:

  stations = PB02, PB08, PB12
  hos_sigma = 500., 1000., 1000.