Joint P and S location.
This commit introduces the possibility of performing a joint P and S
detection and location.

This is quite a large changeset. I'll therefore try to go through it
step by step.

First of all, all this is optional and controlled by the config file
option:

   wave_type = PS

Using 'wave_type = P' or 'wave_type = S' will fall back to classical
P-only or S-only detection and location.

P and S characteristic functions are cross-correlated in the same time
window. This means that the time window has to be large enough to
contain both P and S arrivals. This is a major limitation. Overcoming
it (i.e., allowing for P and S in different time windows) would
introduce a lot more complexity in the code logic, and I'm not sure if
it is worth the effort...

From an implementation point of view, here's the list of main changes:

- NLL gird reading has been moved to a separate module
  ('read_grids.py'). To allow for storing both P and S traveltime
  grids, 'GRD_sta' is now a double dictionary with station as first
  key and wave type as second key, i.e. 'GRD_sta[station][wave]'.

- Creation of summary characteristic function has also been moved to a
  separate module ('summary_cf.py'). Wave type (P or S) is stored in
  each CF trace object as 'channel' attribute.

- '_run_BackProj()' now considers all the possible combinations of
  stations and wave type. This means that, given two stations A and B,
  we have:

    stationA,P-wave  with  stationB,P-wave
    stationA,S-wave  with  stationB,S-wave
    stationA,P-wave  with  stationB,S-wave
    stationA,S-wave  with  stationB,P-wave
    stationA,P-wave  with  stationA,S-wave
    stationB,P-wave  with  stationB,S-wave

  This increases by a factor of six the number of cross-correlations and
  of back-projections to perform.

- 'sta_GRD_Proj()' has been modified to manage, as input parametes, the
  different wave types. Returned arrival times (measured from cross-
  correlation) are now either P or S, and are stored accordingly, by
  '_run_BackProj()', in the 'arrival_times' double dictionary, e.g.
  'arrival_times[station][wave]'.

- 'TrOrig_time()' uses, for the moment, only P arrivals to compute
  origin time. Theoretical arrivals are computed for P and S, but
  probably the S are not correct. I will come back to this, as soon
  as we will have a proper 'Pick()' object, as discussed in issue #4.

- Plotting routines have been modified to manage the different wave
  types. P-wave CF is plotted in blue, S-wave CF in red. I'm open to
  changing these colors. For the moment, I only plot measured CC
  arrival times, using the same color code. Theoretical arrivals are
  not plotted for the same reasons exposed above.
  Another small change in plotting routines is that they will now
  always try to plot the vertical component of the original traces,
  if available.

There are, sure enough, other modifications that I forgot. Take a look
at the code!

All this needs, of course, extensive testing.