Recursive memory: first steps.
Those are the first steps into the implementation of a recursive memory
mechanism to compute MB filter and CF on a window-by-window basis.

With this commit:
- rec_filter, rec_rms and rec_hos have now "memory" variables which are
  kept from call to call;
- a new class, RecursiveMemory(), is defined in bp_types.py, which keeps
  in one place all the "memory" variables;
- a list of RecursiveMemory() objects (whose lenght is the number of
  filter bands) can be optionally passed to MBfilter_CF() to enable the
  memory mechanism. WARNING: this is so far only implemented for
  single-component analysis (i.e., it doesn't work for S-waves obtained
  by Rosenberger filter on horizontals).

TODO: The memory mechanism is not yet implemented in the main code.
The implementation will come in the next commits. So, at this
stage, the main code should run as always.

This commit also optimizes and cleans up the code in lib_rec_hos.c.