Attach station list to config object.
This reduces the number of parameters to pass around to functions.

Also, (optional) data resampling is now done inside "read_traces()" and
we check that all the sampling step "delta" is consistent among all the
traces (otherwhise this throws out an error). The common "delta" value
is stored in the config object.