Use recursive gauss filter instead of convolution
We already have a recurisive gauss filter implementation (in
"lib_rec_cc.c") which can be used in place of the convolution function
of numpy.
The advantage is that this is faster and more suited to recursive
implementation.

When using "numpy.convolve", there are huge differences between the
non-recursive-memory and the recurisve-memory CF.
Using the recursive gauss filter, instead, elminates *almost* all the
differences.

I said "almost", because there is still a problem at window ends, due
to the fact that the gauss filter goes to zero at the end of the window.
This creates a tiny difference with the non-recursive-memory CF for a
few samples, which doesn't seem to be a big deal.
Fixing this problem is a hard task, since the recursive gaussian filter
is a two-pass filter (forwards and backwards), and a simple memory
scheme like the one introduced in the previous commits is not
appropriate.