Station-level parallelization for recurisve memory
When using recursive memory, parallelization is done at station-level,
i.e. every couple of station is analysed in a separate process.

To do that, the interface of sta_GRD_Proj() had to be simplified, in
order to reduce the amount of data to pass between the processes.

Also, plotting is done asynchronously on a different process.

When not using recursive memory, the old time-based parallelization
scheme is used instead. This approach is still 10% faster than the new
one, but, of course, it does not respect the temporal order of
time windows (and, therefore, is not compatible with recursive_memory).