New option: 'ignore_noisy_CF'
If 'ignore_noisy_CF = True' then 'noisy' CF are detected at each time
step with the following algorithm:

- the integral of each CF is computed;
- the CF with the smallest integral is the best one;
- all the CFs whose integral is larger than two times the smallest
  integral are considered noisy and removed from the computation.

Unused CF are plotted with dashed lines.

The number of traces used for a trigger is now reported in the trigger
output (Ntraces).

Theoretical travel times are computed for all the stations.