Option: filter_npoles (replaces filter_strength)
The previous approach (filter_strength) was wrong.

This commit implements a generic bandpass (or highpass) recursive filter
made of a cascade of simple, first-order filters.
The number of poles is the number of filter stages.

By default, the number of poles is two, which corresponds to the
previous version and should provide exactly the same results of the
previous version.

Filter coefficients are now computed by the routine `rec_filter_coeff()`
in `rec_filter.py`.

This commit also improves the testing routine in `rec_filter.py`, which
now shows the filter responses at different frequencies for different
numbers of poles.