Consistently define the decay constant at 5% level
The decay constant is now consistently defined as the time at which the
CF reaches 5% of its initial value.

   decay_constant = 1 - np.exp(np.log(0.05) / decay_samples)

Note: this definition gives a constant which, for decay_samples > ~20,
is roughly three times the one obatined with the previous definition
(1./decay_samples).

For this reason, I changed the value of the `decay_const` option in
`BT_ChileExample.conf` from 1.0 sec to 3.0 sec and `sigma_gauss`
(computed in `summary_cf()`) is now 1./6 of decay_const instead of 1./2.

TODO: the name `decay_const` in the config file is misleading. It would
be better calling it `decay_win`.